+++ NEWS.md
@@ -1,5 +1,6 @@
 ## 1.3.0 unpublished
- * CQLPG-6: Replace regexp by LIKE for <> and ==
+ * CQLPG-6: Replace regexp by LIKE for <> and == for performance
+ * CQLPG-10: Use unaccent instead of regexp for performance
 ## 1.2.1 2017-11-14
  * Replace CASE jsonb_typeof(field) ... by an AND/OR expression (CQLPG-8)
  * For numeric values use -> not ->> to make use of the index. (CQLPG-7 partially)